fix: remove uint8 from collider as it's handled as char by the serializer

+++ common/include/components/Projectile.hpp
@@ -12,7 +12,7 @@
 
 namespace component
 {
-    enum class ProjectileType : uint8_t
+    enum class ProjectileType : int
     {
         PLAYER = 0,
         MOB = 1
